Move older (pre-2019) historical code to historical.legacy
