+++ item/historical/__init__.py
@@ -1,15 +1,18 @@
 from copy import copy
+from functools import lru_cache
 import os
 
 import pandas as pd
-import pint
+import pycountry
 import yaml
 
 from item.common import paths
 from item.remote import OpenKAPSARC, get_sdmx
+from .scripts import T001
+from .scripts.util.managers.dataframe import ColumnName
 
 
-# List of data processing scripts/IPython notebooks
+#: List of data processing Jupyter/IPython notebooks.
 SCRIPTS = [
     'T000',
     'T001',
@@ -22,59 +25,60 @@ SCRIPTS = [
     'T008'
 ]
 
+#:
+MODULES = {
+    1: T001
+}
 
-# Define a registry for tracking of units, and add units appearing in the data.
-ureg = pint.UnitRegistry()
-ureg.define("""
-idx_2005_100 = [index, 2005=100]
-person = [person]
-passenger = person
-TEU = [container]
-TOE = 41.9 GJ
-vehicle = [vehicle]
-yr = year
-""".strip())
+OUTPUT_PATH = paths['data'] / 'historical' / 'output'
+
+#: Non-ISO names appearing in 1 or more data sets; see :meth:`iso_and_region`.
+COUNTRY_NAME = {
+    "Montenegro, Republic of": "Montenegro",
+    "Bosnia-Herzegovina": "Bosnia and Herzegovina",
+    "Korea": "Korea, Republic of",
+    "Serbia, Republic of": "Serbia",
+}
 
 
-def convert_units(row, target_units=[]):
-    """Convert a DataFrame *row* to *target_units*.
+#: Map from ISO 3166 alpha-3 code to region name.
+REGION = {}
+# Populate the map from the regions.yaml file
+with open(paths['data'] / 'model' / 'regions.yaml') as file:
+    for region_name, info in yaml.safe_load(file).items():
+        REGION.update({c: region_name for c in info['countries']})
 
-    The 'Value' and 'Unit' columns of *row* are converted into the first
-    compatible (i.e. same dimensionality) units from *target_units*.
 
-    *target_units* contains zero or more (scaling_factor, target_unit) tuples,
-    where *target_unit* is a pint.Unit and scaling_factor is a float.
+with open(paths['data'] / 'historical' / 'sources.yaml') as f:
+    #:
+    SOURCES = yaml.safe_load(f)
+
+
+def cache_results(id_str, df):
+    """Write *df* to cache in two file formats.
+
+    The files written are:
+
+    - :file:`{id_str}_cleaned_PF.csv`, in long or 'programming-friendly'
+      format.
+    - :file:`{id_str}_cleaned_UF.csv`, in wide or 'user-friendly' format.
     """
-    result = ureg.Quantity(row['Value'], row['Unit'])
-    scaling_factor = 1
+    OUTPUT_PATH.mkdir(exist_ok=True)
 
-    # If no target_units are supplied, this code does not run
-    for scaling_factor, target_unit in target_units:
-        try:
-            result = result.to(target_unit)
-            break
-        except pint.DimensionalityError:
-            # Not a compatible unit for this row
-            continue
-        except Exception as e:
-            print(e, result, target_unit)
-            raise
+    # Long format ('programming friendly view')
+    path = OUTPUT_PATH / f'{id_str}_cleaned_PF.csv'
+    df.to_csv(path, index=False)
+    print(f'Write {path}')
 
-    # Set the resulting values
-    row['Value'] = result.magnitude
-    row['Unit'] = result.units
+    # Pivot to wide format ('user friendly view') and write to CSV
+    columns = [ev.value for ev in ColumnName if ev != ColumnName.VALUE]
+    path = OUTPUT_PATH / f'{id_str}_cleaned_UF.csv'
+    df.set_index(columns) \
+      .unstack(ColumnName.YEAR.value) \
+      .reset_index()\
+      .to_csv(path, index=False)
 
-    # scaling_factor had the value from the preferred unit
-    if scaling_factor != 1:
-        row['Value'] /= scaling_factor
-        row['Unit'] *= scaling_factor
-
-    return row
-
-
-def source_str(id):
-    """Return the canonical string name (e.g. 'T001') for a data source."""
-    return f'T{id:03}' if isinstance(id, int) else id
+    print(f'Write {path}')
 
 
 def fetch_source(id, use_cache=True):
@@ -130,140 +134,99 @@ def input_file(id: int):
     return all_files[-1]
 
 
-def map_values(row, mapping):
-    """Return a mapped value from a DataFrame *row*.
+def process(id):
+    """Process a data set given its *id*.
 
-    *mapping* is a dict with a key '_column'. The value in this column of *row*
-    is used to return a value from *mapping*.
+    Performs the following steps:
+
+    1. Load the data from cache.
+    2. Load a module defining dataset-specific processing steps. This module
+       is in a file named e.g. :file:`T001.py`.
+    3. Call the dataset's (optional) :meth:`check` method. This method receives
+       the input data frame as an argument, and can make one or more assertions
+       to ensure the data is in the expected format.
+    4. Drop columns in the dataset's (optional) :data:`DROP_COLUMNS`
+       :class:`list`.
+    5. Call the dataset's (required) :meth:`process` method. This method
+       receives the data frame from step (4), and performs any necessary
+       processing.
+    6. Assigns ISO 3166 alpha-3 codes and the iTEM region based on a column
+       containing country names.
+    7. Orders columns.
+    8. Outputs data to two files.
+
     """
-    column = mapping['_column']
-    return mapping[row[column]]
+    # Creating the dataframe and viewing the data
 
+    # Creating a dataframe from the csv data
+    id_str = source_str(id)
+    path = paths['data'] / 'historical' / 'input' / f'{id_str}_input.csv'
+    df = pd.read_csv(path)
 
-def preferred_units(info):
-    """Return a list of (scaling_factor, unit) tuples.
+    # Get the module for this data set
+    dataset_module = MODULES[1]
 
-    *info['preferred_units']* may contain a list of one or more strings
-    describing units with optional scaling factors.
-    """
-    result = []
+    try:
+        # Check that the input data is of the form required by the script
+        dataset_module.check(df)
+    except AttributeError:
+        print('No pre-processing checks to perform')
+    except AssertionError as e:
+        print(f'Input data is invalid: {e}')
 
-    for unit in info.get('preferred_units', []):
-        try:
-            result.append[(1, ureg.parse_units(unit))]
-        except ValueError:
-            # *unit* has a scaling factor, e.g. 10³ km. Split this to the
-            # scaling factor and the pure unit (e.g. 'km').
-            qty = ureg.parse_expression(unit)
-            result.append[(qty.magnitude, qty.units)]
-        except Exception as e:
-            print(e, unit)
-            raise
+    try:
+        # Remove unnecessary columns
+        df.drop(columns=dataset_module.DROP_COLUMNS, inplace=True)
+        print('Drop {len(dataset_module.DROP_COLUMNS)} extra column(s)')
+    except AttributeError:
+        # No variable DROP_COLUMNS in dataset_module
+        print(f'No columns to drop for {id_str}')
 
-    return result
+    # Call the dataset-specific processing
+    df = dataset_module.process(df)
+    print(f'{len(df)} observations')
 
+    # Perform common cleaning tasks
 
-# Function of conversion phase 1
-def conversion_layer1(df, top_dict={}):
-    """Convert *df* to a standard format.
+    # Assign ISO 3166 alpha-3 codes and iTEM regions from a country name column
+    column = 'Country'  # TODO read this name from dataset_module
 
-    *top_dict* may contain each of the following keys, which are processed in
-    the order that they appear here. See mapping_conv_phase1.yaml for concrete
-    examples.
+    # Use pandas.Series.apply() to apply the same function to each entry in
+    # the given column
+    df = pd.concat([df, df[column].apply(iso_and_region)], axis=1)
 
-    - 'rename': a mapping from original column names to final column names.
-      All other columns are dropped.
-    - 'add': a mapping from column names to values which are used to fill all
-      rows. The value may itself be a mapping; in this case, it is passed to
-      :meth:`map_values` to look up the values in the 'add' column from values
-      in the '_column' column.
-    - 'replace': a mapping from column names to a replacement mapping that is
-      applied to that column.
-    - 'preferred_units': a list of target units
+    # Reordering the columns
+    #
+    # Rule: The columns should follow the order established in the latest
+    # template
+    df = df.assign(**{ColumnName.ID.value: id_str}) \
+           .reindex(columns=[ev.value for ev in ColumnName])
 
-    Parameters
-    ----------
-    df : pd.DataFrame
-        Raw data set loaded from IK2 Open Data.
-    top_dict : dict, optional
-        Dictionary of mapping rules.
-    """
+    cache_results(id_str, df)
 
-    # Rename existing columns to preferred dimensions, dropping others
-    rename = top_dict.get('rename', {})
-    df = df.take(rename.keys(), axis=1) \
-           .rename(columns=rename)
-
-    # Add columns that are not included and mapping with pre-defined rules
-    for col, value in top_dict.get('add', {}).items():
-        if '_depend' in value:
-            df[col] = df.apply(map_values, axis=1, args=(value,))
-        else:
-            df[col] = value
-
-    # Replace values in fields
-    for col, mapping in top_dict.get('replace', {}).items():
-        df[col].replace(mapping, inplace=True)
-
-    # Replace '-' with ' '. pint interprets '-' to mean subtraction.
-    df['Unit'] = df['Unit'].str.replace('-', ' ')
-
-    # Convert units
-    df = df.apply(convert_units, axis=1, args=(preferred_units(top_dict),))
-
+    # Return the data for use by other code
     return df
 
 
-with open(paths['data'] / 'historical' / 'sources.yaml') as f:
-    SOURCES = yaml.safe_load(f)
+@lru_cache()
+def iso_and_region(name):
+    """Return (ISO 3166 alpha-3 code, iTEM region) for a country *name*."""
+    # lru_cache() ensures this function call is as fast as a dictionary lookup
+    # after the first time each country name is seen
+
+    # Maybe map a known, non-standard value to a standard value
+    name = COUNTRY_NAME.get(name, name)
+
+    # Use pycountry's built-in, case-insensitive lookup on all fields including
+    # name, official_name, and common_name
+    alpha_3 = pycountry.countries.lookup(name).alpha_3
+
+    # Look up the region, construct a Series, and return
+    return pd.Series(
+        [alpha_3, REGION.get(alpha_3, 'N/A')],
+        index=[ColumnName.ISO_CODE.value, ColumnName.ITEM_REGION.value])
 
 
-def main(output_file, use_cache):
-    """Convert the input datasets.
-
-    A single dataframe with the output is written to *output_file*.
-    If *use_cache* is True, local files are used before querying the API.
-    """
-    # Get the configuration
-    # TODO load these from all the files appearing in one directory, so that
-    #      each data set can be specified in a separate file
-    ds_info_path = paths['data'] / 'historical' / 'mapping_conv_phase1.yaml'
-    all_ds_info = yaml.safe_load(open(ds_info_path))
-
-    # Access the OpenKAPSARC API
-    ok = OpenKAPSARC()
-
-    list_of_df = []
-
-    for ds_info in all_ds_info:
-        if use_cache:
-            # Locate a previously-cached CSV file
-            cache_path = (paths['historical'] /
-                          ds_info['uid']).with_suffix('.csv')
-            try:
-                df = pd.read_csv(cache_path, sep=';')
-                cache_miss = False
-            except FileNotFoundError:
-                cache_miss = True
-
-        if not use_cache or cache_miss:
-            # Retrieve the data from the OpenKAPSARC datahub
-            df = ok.table(ds_info['id'])
-
-        # Process the data
-        df = conversion_layer1(df, ds_info)
-
-        # Add a source annotation
-        df['Source'] = f"OpenKAPSARC:{ds_info['uid']}"
-
-        # Store
-        list_of_df.append(df)
-
-    # Combine to single dataframe
-    df_output = pd.concat(list_of_df, sort=False, ignore_index=True)
-    df_output = df_output[["Region", "Variable", "Unit", "Mode", "Technology",
-                           "Fuel", "Year", "Value", "Source"]]
-
-    df_output.to_csv(output_file, index=False)
-
-    return df_output
+def source_str(id):
+    """Return the canonical string name (e.g. 'T001') for a data source."""
+    return f'T{id:03}' if isinstance(id, int) else id
