+++ item/tests/test_historical.py
@@ -5,8 +5,13 @@ import pytest
 
 import item
 from item.common import paths
-from item.historical import SCRIPTS, fetch_source, input_file, source_str
-from item.historical.scripts import process
+from item.historical import (
+    SCRIPTS,
+    fetch_source,
+    input_file,
+    process,
+    source_str,
+)
 from item.historical.diagnostic import coverage
 from item.historical.util import run_notebook
